Restructure PurePursuitCommand to allow custom implementations
